fix comment

+++ enginetest/queries/queries.go
@@ -1022,14 +1022,15 @@ var QueryTests = []QueryTest{
 		Expected: []sql.Row{{"1.0"}, {"2.5"}},
 	},
 	{
+		// The SortFields does not match between prepared and non-prepared nodes.
 		SkipPrepared: true,
 		Query:        `SELECT column_0 FROM (values row('1.5',2+2), row(floor(1.5),concat("a","b"))) a order by 1;`,
 		Expected:     []sql.Row{{"1"}, {"1.5"}},
 	},
 	{
-		// it's float '1' and '1.5' but instead it should have '1.0' and '1.5' decimal results
+		// the result on sql shell are '1' and '1.5' but instead it should have decimal values of '1.0' and '1.5'
 		Query:    `SELECT column_0 FROM (values row(1.5,2+2), row(floor(1.5),concat("a","b"))) a order by 1;`,
-		Expected: []sql.Row{{1.0}, {1.5}},
+		Expected: []sql.Row{{float64(1)}, {1.5}},
 	},
 	{
 		Query: `SELECT FORMAT(val, 2) FROM
